refactor: clean old carousel section
