+++ app/layout/home/banner/index.ts
