--- app/layout/home/carousel-section/constants.ts
@@ -1,27 +1,0 @@
-import CarouselImage1 from 'images/home-carousel/carousel_01.png';
-import CarouselImage2 from 'images/home-carousel/carousel_02.png';
-import CarouselImage3 from 'images/home-carousel/carousel_03.png';
-import CarouselImage4 from 'images/home-carousel/carousel_04.png';
-
-export const IMAGES = [
-  {
-    alt: 'project-features',
-    id: 1,
-    src: CarouselImage1,
-  },
-  {
-    alt: 'project-features',
-    id: 2,
-    src: CarouselImage2,
-  },
-  {
-    alt: 'project-analysis',
-    id: 3,
-    src: CarouselImage3,
-  },
-  {
-    alt: 'project-solutions',
-    id: 4,
-    src: CarouselImage4,
-  },
-];
