+++ app/package.json
@@ -20,6 +20,7 @@
     "@dnd-kit/modifiers": "^2.0.0",
     "@dnd-kit/sortable": "^3.0.1",
     "@dnd-kit/utilities": "^2.0.0",
+    "@egjs/flicking-plugins": "^4.2.2",
     "@egjs/react-flicking": "^4.2.2",
     "@hapi/iron": "^6.0.0",
     "@math.gl/web-mercator": "^3.3.2",
